# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tinder App API",
        "description": "API for Tinder-like dating app MVP",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/v1/people/{id}/like": {
            "post": {
                "tags": [
                    "Interactions"
                ],
                "summary": "Like a person",
                "operationId": "c8b7fba08d531bdba51f5432e983afe0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Like recorded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "SUCCESS"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Like recorded successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "person_id": {
                                                    "type": "integer",
                                                    "example": 57
                                                },
                                                "action": {
                                                    "type": "string",
                                                    "example": "like"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "recorded"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Duplicate like",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "DUPLICATE_LIKE"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You have already liked this person."
                                        },
                                        "data": {
                                            "properties": {
                                                "person_id": {
                                                    "type": "integer",
                                                    "example": 57
                                                },
                                                "action": {
                                                    "type": "string",
                                                    "example": "like"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "duplicate"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "UserIdHeader": []
                    }
                ]
            }
        },
        "/v1/people/{id}/dislike": {
            "post": {
                "tags": [
                    "Interactions"
                ],
                "summary": "Dislike a person",
                "operationId": "e8c0284147b494814e63a5a0423bf0cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Dislike recorded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "SUCCESS"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dislike recorded successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "person_id": {
                                                    "type": "integer",
                                                    "example": 58
                                                },
                                                "action": {
                                                    "type": "string",
                                                    "example": "dislike"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "recorded"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "UserIdHeader": []
                    }
                ]
            }
        },
        "/v1/likes": {
            "get": {
                "tags": [
                    "Interactions"
                ],
                "summary": "Get list of liked people",
                "operationId": "9a0aa70e086b3ecf978313b82b15782c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of liked people",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "SUCCESS"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of liked people"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Person"
                                                    }
                                                },
                                                "meta": {
                                                    "properties": {
                                                        "page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 20
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "total_pages": {
                                                            "type": "integer",
                                                            "example": 1
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "UserIdHeader": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Person": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 116
                    },
                    "name": {
                        "type": "string",
                        "example": "Ayu"
                    },
                    "age": {
                        "type": "integer",
                        "example": 26
                    },
                    "location": {
                        "type": "string",
                        "example": "Jakarta"
                    },
                    "pictures": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "example": "https://.../ayu1.jpg"
                                },
                                "order": {
                                    "type": "integer",
                                    "example": 1
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "UserIdHeader": {
                "type": "apiKey",
                "description": "User ID for testing (UUID format)",
                "name": "X-User-Id",
                "in": "header"
            }
        }
    },
...
}